services/common/requirements.txt
fastapi>=0.110.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
